Enable displaying full-size images
The gallery is populated with thumbnails by default. If no thumbnail
exists for a given image, but the full-size version exists, then the
full-size version is inserted in place of the thumbnail.

If a full-size version of an image exists, its thumbnail view will be
clickable. Clicking the thumbnail displays the image scaled to fit the
screen as an overlay. Clicking anywhere on the overlay hides it again.
Thumbnails which do not have a full-size image are not clickable.