Improve retrieval of version details from database
This commit requires corresponding changes in the database.

Additional error handling is needed for the scenario where fallback
version number cannot be retrieved from the conf file.

Constructed hrefs should be tested by the code and plain text used
instead of <a> tags if the referenced file doesn't exist (TODO).